% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.meta.R
\name{average.meta.freqs}
\alias{average.meta.freqs}
\title{Average variant frequencies}
\usage{
average.meta.freqs(df)
}
\arguments{
\item{df}{Merged data.frame of two or more cohorts \(see \code{Details}\)}
}
\value{
data.frame
}
\description{
Compute weighted average of variant frequencies for germline-somatic association stats
}
\details{
\code{df} should be a data.frame generated by \link{merge.assoc.stats}
}
\seealso{
\link{load.assoc.stats.single}, \link{merge.assoc.stats}
}
