% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.basic.R
\name{compress.ad.matrix}
\alias{compress.ad.matrix}
\title{Compress allele dosage matrix}
\usage{
compress.ad.matrix(ad.df, action, elig.controls = NULL)
}
\arguments{
\item{ad.df}{An allele dosage matrix imported by \link{load.ad.matrix}}

\item{action}{Action to apply to query rows; see \code{Details}. Required.}

\item{elig.controls}{Sample IDs eligible to be reported as 0 [default: all samples]}
}
\value{
numeric or logical vector, depending on \code{action}
}
\description{
Compress an allele dosage matrix based on a specified compression function
}
\details{
Recognized values for \code{action} include:
\itemize{
\item \code{"verbose"} : return the full query matrix [default]
\item \code{"any"} : return numeric indicator if any of the query rows are non-zero
for each sample
\item \code{"all"} : return numeric indicator if all of the query rows are non-zero
for each sample
\item \code{"sum"} : return the sum of allele dosages for all query rows per sample
\item \code{"max"} : return the max of allele dosages for all query rows per sample
}
}
\seealso{
\link{load.ad.matrix}, \link{query.ad.matrix}
}
