% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.basic.R
\name{germline.somatic.assoc}
\alias{germline.somatic.assoc}
\title{Germline-somatic association test}
\usage{
germline.somatic.assoc(
  y.vals,
  x.vals,
  meta,
  gqs = NULL,
  firth.fallback = TRUE,
  strict.fallback = TRUE,
  nonstrict.se.tolerance = 10,
  firth.always = FALSE,
  custom.covariates = c(),
  multiPop.min.ac = 10,
  multiPop.min.freq = 0.01,
  logistf.maxiter = 1000,
  model.suffix = "",
  only.return.freqs = FALSE
)
}
\arguments{
\item{y.vals}{Numeric vector indicating presence or absence of somatic endpoint}

\item{x.vals}{Numeric vector of germline values to test}

\item{meta}{Metadata for all samples as loaded by \link{load.patient.metadata}}

\item{gqs}{Two-element list of mean ref & alt genotype qualities per sample.
If provided, will be included as covariates in association model.}

\item{firth.fallback}{Attempt to use Firth bias-reduced logistic regression when
traditional logistic regression fails to converge or dataset is quasi-separable
[default: TRUE]}

\item{strict.fallback}{Implement Firth regression if a standard logit model returns
any errors or warnings. Setting this to \code{FALSE} will only default to Firth
regression if logit returns any errors or if the standard error of the genotype
coefficient exceeds \code{nonstrict.se.tolerance} [default: TRUE]}

\item{nonstrict.se.tolerance}{If \code{strict.fallback} is \code{FALSE}, only use Firth
regression if a standard logit model produces a genotype effect standard error
exceeding this value [default: 10]}

\item{firth.always}{Always use Firth regression [default: FALSE]}

\item{custom.covariates}{Character vector of columns from \code{meta} to include
as covariates in the regression model. See \code{Details}.}

\item{multiPop.min.ac}{Minimum allele count to include non-European populations}

\item{multiPop.min.freq}{Minimum variant frequency to include non-European populations}

\item{logistf.maxiter}{Maximum number of iterations for \code{logistf()} [default: 1000]}

\item{model.suffix}{String to append to the end of the "model" indicator in
the returned vector of results [default: append nothing]}

\item{only.return.freqs}{If set to \code{TRUE}, no association will be conducted
and only frequencies will be reported. All association statistics will be
reported as \code{NA} in the returned vector.}
}
\value{
Vector of test results
}
\description{
Run association test for a single germline-somatic interaction
}
\details{
By default, the following covariates will be included:
\itemize{
\item Age, sex, and an interaction term for age by sex
\item Top ten principal components \(PCs\)
\item Tumor purity
}

Any other covariates to be included must be specified in \code{custom.covariates}
}
\seealso{
\link{load.patient.metadata}, \link[logistf:logistf]{logistf::logistf}
}
