% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.meta.R
\name{germline.somatic.meta.analysis}
\alias{germline.somatic.meta.analysis}
\title{Meta-analyze effect sizes}
\usage{
germline.somatic.meta.analysis(df, model = "FE")
}
\arguments{
\item{df}{Merged data.frame of two or more cohorts \(see \code{Details}\)}

\item{model}{Specify model architecture \(see \code{Details}\)}
}
\value{
data.frame
}
\description{
Meta-analyze effect sizes of all germline-somatic associations
}
\details{
\code{df} should be a data.frame generated by \link{merge.assoc.stats}

See \link[metafor:rma.uni]{metafor::rma} for eligible values of \code{model}
}
\seealso{
\link{load.assoc.stats.single}, \link{merge.assoc.stats}, \link{ivw.meta.single}
}
