% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.genomic.R
\name{infer.germline.position.from.id}
\alias{infer.germline.position.from.id}
\title{Infer germline position}
\usage{
infer.germline.position.from.id(vid, extract = "position")
}
\arguments{
\item{vid}{Germline variant ID}

\item{extract}{Value to extract, either "position" or "chromosome" [default: "position"]}
}
\value{
Single value (either numeric or character) depending on \code{extract}
}
\description{
Extract the position of a germline variant based on its variant ID
}
\details{
Note that this function assumes the variant adheres to the
ID format of chrom_pos_ref_alt used throughout this study
}
