% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.patient.metadata}
\alias{load.patient.metadata}
\title{Load Patient Metadata}
\usage{
load.patient.metadata(
  file,
  fill.missing = NA,
  fill.columns = NULL,
  deduplicate = FALSE
)
}
\arguments{
\item{file}{Path to input patient metadata .tsv}

\item{fill.missing}{Behavior for handling missing values. See \code{Details}. [default: NA]}

\item{fill.columns}{Specify in which columns missing values should be filled.}

\item{deduplicate}{Deduplicate based on patient ID / first column [default: FALSE]
See \code{Details}. [default: NULL]}
}
\value{
data.frame of patient metadata
}
\description{
Load a standard patient metadata file for a single cohort
}
\details{
See \link{impute.missing.values} for information on \code{fill.missing} and \code{fill.columns}.
To disable missing value imputation, provide \code{fill.missing} = \code{NA}.
}
\seealso{
\link{impute.missing.values}
}
