% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{mode}
\alias{mode}
\title{Arithmetic Mode}
\usage{
mode(values, break.ties = "all", seed = 2022)
}
\arguments{
\item{values}{Vector of values}

\item{break.ties}{Specify how ties should be broken. See \code{Details}. [default: "all"]}

\item{seed}{Random seed to break ties, if needed. [default: 2022]}
}
\value{
mode(s) of \code{values}
}
\description{
Compute the arithmetic mode of a vector
}
\details{
Recognized values for \code{break.ties} include:
\itemize{
\item \code{all} : return all tied modes
\item \code{first} : return the first mode value encountered
\item \code{last} : return the last mode value encountered
\item \code{random} : return a randomly selected mode value
}
}
