% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.basic.R
\name{query.ad.matrix}
\alias{query.ad.matrix}
\title{Query allele dosage matrix}
\usage{
query.ad.matrix(
  ad,
  vids,
  elig.controls = NULL,
  action = "verbose",
  missing.vid.fill = NA,
  na.behavior = "all",
  na.frac = 1
)
}
\arguments{
\item{ad}{One or more allele dosage matrix(es) as imported by
\link{load.ad.matrix}. See \code{Details}}

\item{vids}{Variant IDs to query}

\item{elig.controls}{Sample IDs eligible to be reported as 0 [default: all samples]}

\item{action}{Action to apply to query rows; see \code{Details}}

\item{missing.vid.fill}{Value to fill for all samples if a specified variant
ID from \code{vids} is not present in \code{ad} [default: NA]}

\item{na.behavior}{Specify how \code{NA} entries in \code{ad.df} should be treated when
compressing. See \link{compress.ad.matrix} for more information. [default: "all"]}

\item{na.frac}{Fraction of \code{NA} entries allowed before failing a sample. [default: 1]
See \link{compress.ad.matrix} for more information.}
}
\value{
numeric vector or data.frame, depending on \code{action}
}
\description{
Subsets and optionally compresses an allele dosage matrix
}
\details{
The \code{ad} argument accepts either a single data.frame or a list
of data.frames. If a list is provided, each matrix will be queried individually
and their results will be concatenated.

Recognized values for \code{action} include:
\itemize{
\item \code{"verbose"} : return the full query matrix [default]
\item \code{"any"} : return numeric indicator if any of the query rows are non-zero
for each sample
\item \code{"all"} : return numeric indicator if all of the query rows are non-zero
for each sample
\item \code{"sum"} : return the sum of allele dosages for all query rows per sample
\item \code{"max"} : return the max of allele dosages for all query rows per sample
\item \code{"mean"} : return the mean of allele dosages for all query rows per sample
}
}
\seealso{
\link{load.ad.matrix}, \link{compress.ad.matrix}
}
