% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.basic.R
\name{query.gq.matrix}
\alias{query.gq.matrix}
\title{Query genotype quality matrix}
\usage{
query.gq.matrix(gq, ad, vids = NULL, fill.mean = TRUE)
}
\arguments{
\item{gq}{One or more matrixes of genotype qualities imported by
\link{load.ad.matrix}. See \code{Details}.}

\item{ad}{One or more allele dosage matrixes imported by \link{load.ad.matrix}.
Order must match \code{gq}. See \code{Details}.}

\item{vids}{Variant IDs to query [default: use all VIDs in \code{ad.matrix}]}

\item{fill.mean}{Fill missing mean GQs with cohort-wide mean on a
genotype-specific basis [default: TRUE]}
}
\value{
list of two numeric vectors, one for reference genotypes and one for
non-reference genotypes
}
\description{
Extract and compute genotype-conditional mean GQ values per sample
from a genotype quality matrix
}
\details{
The \code{gq} argument accepts either a single data.frame or a list
of data.frames. If a list is provided, each matrix will be queried individually
and their results will be concatenated.
}
\seealso{
\link{load.ad.matrix}, \link{query.ad.matrix}
}
