% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{scaled.swarm}
\alias{scaled.swarm}
\title{Scale-Aware Beeswarm Cluster}
\usage{
scaled.swarm(
  values,
  colors,
  group.names = NULL,
  sep.wex = 0.05,
  pch = 19,
  pt.cex = 0.2,
  y.title = NULL,
  y.title.line = 0.5,
  y.axis.at = NULL,
  y.axis.labels = NULL,
  parmar = c(1, 2.5, 0.25, 0.25)
)
}
\arguments{
\item{values}{List of numeric vectors of values to plot}

\item{colors}{Vector of colors for the list elements in \code{values}}

\item{group.names}{(Optional) group names to assign to each list element in \code{values}}

\item{sep.wex}{Relative width scalar for whitespace on the X-axis
between groups [default: 0.05]}

\item{pch}{Value passed to \code{beeswarm}}

\item{pt.cex}{Value of \code{cex} passed to \code{beeswarm}}

\item{y.title}{Title of Y-axis}

\item{y.title.line}{Value of \code{line} for \code{y.title}}

\item{y.axis.at}{Custom Y-axis tick positions, if desired}

\item{y.axis.labels}{Custom Y-axis tick labels, if desired}

\item{parmar}{Margin values passed to par()}
}
\description{
Plot a set of beeswarm distributions scaled proportional to set size
}
\details{
If \code{values} is supplied as a named list, those names will be used as
\code{group.names} unless \code{group.names} is explicitly specified. Otherwise, \code{group.names}
will be set to the ordinal value of each group in \code{values}.
}
\seealso{
\link{scaled.bars}, \link{clean.axis}
}
