% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.genomic.R
\name{simplify.vids}
\alias{simplify.vids}
\title{Simplify allele dosage variant IDs}
\usage{
simplify.vids(ad.df)
}
\arguments{
\item{ad.df}{Allele dosage matrix as read by \code{\link[RLCtools:query.ad.matrix]{RLCtools::query.ad.matrix()}}}
}
\value{
Allele dosage matrix with simplified variant IDs.
}
\description{
Simplify variant IDs for a single allele dosage matrix
}
\details{
Note that this function assumes the variant IDs adhere to the
ID format of chrom_pos_ref_alt used throughout this study
}
