% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.basic.R
\name{query.gq.matrix}
\alias{query.gq.matrix}
\title{Query genotype quality matrix}
\usage{
query.gq.matrix(gq.matrix, ad.matrix, vids = NULL, fill.mean = TRUE)
}
\arguments{
\item{gq.matrix}{Matrix of genotype qualities imported by \link{load.ad.matrix}}

\item{ad.matrix}{Allele dosage matrix imported by \link{load.ad.matrix}}

\item{vids}{Variant IDs to query [default: use all VIDs in \code{ad.matrix}]}

\item{fill.mean}{Fill missing mean GQs with cohort-wide mean on a
genotype-specific basis [default: TRUE]}
}
\value{
list of two numeric vectors, one for reference genotypes and one for
non-reference genotypes
}
\description{
Extract and compute genotype-conditional mean GQ values per sample
from a genotype quality matrix
}
\seealso{
\link{load.ad.matrix}, \link{query.ad.matrix}
}
